Bail out of ENTRYPOINT approach from now, incompatible with VSCode.
ENTRYPOINT creates a user matching the host user during the initial
docker run (when container starts). This works nicely, but VSCode
tries to be smart (and inspects /etc/password) before starting the
container; this is incompatible.

@TODO Find out if we can make it to work.

The fall-back is the old approach where we create a few user IDs
starting from 1000, in the assumption the host user IDs are like
that also.